Removed all CPL parts, and changed standby PSU.
The Holtek was a SOT-89 that had a completely different pinout to
every other SOT-89 3.3V LDO out there. To avoid confusion I
wanted to replace it, but the only other basic JLCPCB parts
were the AMS1117 (huge quiescent current) and something only
rated to 7V input voltage (this LDO gets the full VDC in).

So, I used a TO-92 footprint and picked an LP2950N.

This, however, needs some ESR on the output cap, so I made
that a 2.2uF electrolytic.

I also removed the 2x 10uF output caps on the K7805 and set
just one 22uF as specified in the datasheet.