Re-organised top sheet.
Lots of parts just changed their component designator. I had
to patch up the PCB manually, but now it imports from the schematic
OK.

Also adds two PSU rails for the RTC (so you can pick DS1307
which is a 5V part, or MCP7940N which is a 3.3V part).